requirements.txt
Flask==2.1.2
joblib==1.1.0

numpy==1.21.6
Pillow==9.1.0

tensorflow==2.8.0

pandas~=1.3.5
scikit-learn~=0.22.1